Split tranaction to its own module and rename it to UnreliableTransaction
